Move maps to components folder and delete patterns folder
